+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ### Features
 - Server shop now supports modded items defined in `prices.json`
-
+- Added a config option to disable the server shop entirely